migration updated

+++ src/main/resources/db/changelog/changeset/V007_account-service_acc_numbers_seq.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS account_numbers_sequence(
 type VARCHAR(32) NOT NULL,
 counter BIGINT NOT NULL,
-version BIGINT
+version int
 );
